msm: mdss: Organize regulator supplies used by DSI controller
Regulator supplies in the DSI driver are used for two different
purposes. One set of regulators is used as a supply for the DSI
controller while another set is used to supply voltage to the
panel attached to the DSI controller. To support advanced power
saving features, it is required to be able to power down just the
DSI controller while keeping the panel on. To enable this,
organize the supplies into logical power modules.

Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>