msm: mdss: do not call dsi soft reset for idle power collapse
The DSI branch blocks needs to be on in order to issue a DSI software
reset. Commit e8c272ba5c3f23e6cece61bde7834cab7922a1d4 introduced the
change which moved the DSI software reset prior to enabling the DSI
branch clocks in order to support the idle power collapse feature.
This is not correct since a soft reset is not required at all when
resuming from an idle power collapsed state. Fix this by moving the
software reset to the DSI on sequence.

Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>